Created A Branch that does not rely on separate game service
Although the game service is good in theory, I feel it is slightly more complex.
From personal experience I find that having a single project that generates all tables needed in the DB whether fullstack or api is easier to manage and establish relationships with. I feel a cohesive project will make it easier to manage who owns what games later as well.  Hopefully this does not disappoint, this project is tough, and am just trying to simplify for all of us. We can discuss more later. Will upload ERD generated as well to shared folder.